Fixed data type, added additional check

+++ cvat/apps/dataset_manager/formats/cvat.py
@@ -523,7 +523,7 @@ def load(file_object, annotations):
                 shape['z_order'] = int(el.attrib.get('z_order', 0))
 
                 if el.tag == 'box':
-                    shape['rotation'] = int(el.attrib.get('rotation', 0))
+                    shape['rotation'] = float(el.attrib.get('rotation', 0))
                     shape['points'].append(el.attrib['xtl'])
                     shape['points'].append(el.attrib['ytl'])
                     shape['points'].append(el.attrib['xbr'])
